Always send No Data category when creating Categoricals
Adding a categorical variable that doesn't have missing values to a dataset with an exclusion applied failed with "No such category id: '-1'". This is a gap in the server's logic, but it is solvable by always creating categoricals with "No Data" included. Most of the changes here are test updates; however, this change exposed two buglets in crtabs with respect to useNA="ifany". One is fixed here; the other is deferred.